requirements.txt
Flask
gunicorn
werkzeug
pandas
openpyxl
psycopg2-binary